Use ami that receives config information;  moved files around to hide old methodology;  made ray code non-Baseline specific
