+++ autoscaler/ray_baseline_aws.yaml
@@ -19,11 +19,11 @@ auth:
 
 head_node:
   InstanceType: p2.xlarge
-  ImageId: ami-09de87c68cf55d9ed   # This is the eval3.5 baseline
+  ImageId:  ami-09fcc141b05b116dc    # This is the eval3.5 baseline
 
 worker_nodes:
   InstanceType: p2.xlarge
-  ImageId: ami-09de87c68cf55d9ed   # This is also the eval3.5 baseline
+  ImageId: ami-09fcc141b05b116dc   # This is also the eval3.5 baseline
 
 # If needed, can copy files to remote machines.  However, they should be configured to run from a script,
 # so this should not be needed
@@ -31,10 +31,10 @@ file_mounts: {
 }
 
 setup_commands:
-  - echo 'export PATH="$HOME/anaconda3/envs/tensorflow_p36/bin:$PATH"' >> ~/.bashrc
+    - echo 'export PATH="$HOME/anaconda3/envs/tensorflow_p36/bin:$PATH"' >> ~/.bashrc
   # Can't seem to use the default which uses a wheel, which doesn't work on AWS Ubuntu ????
-  - source activate pytorch_latest_p37
-  - pip install -U ray==1.2.0
+  # - source activate pytorch_latest_p37
+  # - pip install -U ray==1.2.0
 
 # Command to start ray on the head node. You don't need to change this.
 head_start_ray_commands:
