+++ aws_scripts/mess_runall.sh
