+++ configs/baseline_local.ini
@@ -1,5 +1,5 @@
 #
-# Config for running Ray for Baseline locally
+# Execution configuration for running Ray for Baseline locally
 #
 [MCS]
 run_script = /home/clark/work/mcs/eval3.5/baseline/gravity.sh
