+++ pipeline/cora_task_runner.py
@@ -5,20 +5,20 @@ import threading
 
 from pipeline import logger
 from pipeline import util
+from pipeline.cora_singletask import CoraSingleTask
 from pipeline.mcs_test_runner import McsTestRunner
-from pipeline.opics_config_change import OpicsConfigChange
-from pipeline.opics_singletask import OpicsSingleTask
-from pipeline.xserver_check import XServerCheck
 from pipeline.xserver_startup import XServerStartup
 
-# Uncomment one of the following.  single is for testing;  the other
-# is for intphys
-# TASK_FILE_PATH = "tasks_delta_echo_foxtrot.txt"
+# Uncomment one of the following.  single is for testing;
 # TASK_FILE_PATH = "scenes_single_scene.txt"
-TASK_FILE_PATH = "scenes_juliett.txt"
+TASK_FILE_PATH = "../scenes_juliett.txt"
+
+# Config file location
+config_on_local = "mcs_config_cora_level2.ini"
+remote_config_file_directory = "/home/ubuntu/workspace/GenPRAM.jl/GenAgent/omg/"
 
 
-class OpicsRunTasks:
+class CoraRunTasks:
 
     def __init__(self):
         self.available_machines = []
@@ -49,7 +49,7 @@ class OpicsRunTasks:
             task_file = task_files_list.pop(0)
             lock.release()
 
-            singleTask = OpicsSingleTask(machine_dns, task_file, threadlog)
+            singleTask = CoraSingleTask(machine_dns, task_file, threadlog)
             return_code = singleTask.process()
 
             if return_code > 0:
@@ -74,13 +74,16 @@ class OpicsRunTasks:
         self.log.info(f"Tasks {task_files_list}")
 
     def run_tasks(self):
+        """ Run the tasks (you must call get_tasks first) on all the machine
+        (you must call get_machines first)"""
+
         # Create a thread for each machine
         threads = []
         for machine in self.available_machines:
-            processThread = threading.Thread(target=self.run_thread_on_ec2_machine,
-                                             args=(machine,))
-            processThread.start()
-            threads.append(processThread)
+            process_thread = threading.Thread(target=self.run_thread_on_ec2_machine,
+                                              args=(machine,))
+            process_thread.start()
+            threads.append(process_thread)
 
         # Wait for them all to finish
         for thread in threads:
@@ -89,17 +92,17 @@ class OpicsRunTasks:
         self.log.info("Ending runtasks")
 
     def get_machines(self):
-        self.available_machines = util.get_aws_machines(tag_name='ta1', tag_value='opics')
+        self.available_machines = util.get_aws_machines(tag_name='ta1', tag_value='cora')
         self.log.info(f"Number of machines {len(self.available_machines)}")
         self.log.info(f"Machines available:  {self.available_machines}")
 
     def run_xstartup(self):
         ''' Start X Server on all the machines.  Note:  Not parallelized'''
         for machine in self.available_machines:
-            cmd = "sudo /usr/bin/Xorg :0"
+            cmd = "/home/ubuntu/start_cora_docker.sh"
             return_code = util.shell_run_background_remote(machine, cmd, self.log)
-            if not return_code == 0:
-                self.log.warn(f"Error starting x on {machine}")
+            self.log.debug(f"Tried to start docker on CORA machine {machine}" +
+                           f" Result: {return_code}")
 
     def kill_and_restartX(self):
         for machine in self.available_machines:
@@ -108,23 +111,29 @@ class OpicsRunTasks:
 
     def change_mcs_config(self):
         for machine in self.available_machines:
-            config_change = OpicsConfigChange(machine, self.log)
-            config_change.process()
+            return_code = util.copy_file_to_aws(machine, config_on_local,
+                                                self.log,
+                                                remote_config_file_directory)
+            if not return_code == 0:
+                self.log.warn(f"Error copying new config {return_code}")
 
     def run_check_xorg(self):
         ''' Check X Server on all the machines.  Note:  Not parallelized'''
         for machine in self.available_machines:
-            xserver_check = XServerCheck(machine, self.log)
-            xserver_check.process()
+            cmd = "docker exec `docker ps -a | grep cora | awk '{print $1}'` ps auxwww | grep Xorg"
+            return_code = util.shell_run_command_remote(machine, cmd, self.log)
+            self.log.debug(f"X Status on remote machine {machine}" +
+                           f" Result: {return_code}")
 
     def run_test(self):
+        self.available_machines = util.get_aws_machines()
         for machine in self.available_machines:
             test_runner = McsTestRunner(machine, self.log)
             test_runner.process()
 
 
 if __name__ == '__main__':
-    run_tasks = OpicsRunTasks()
+    run_tasks = CoraRunTasks()
     run_tasks.get_machines()
     run_tasks.get_tasks()
 
@@ -133,6 +142,6 @@ if __name__ == '__main__':
     # run_tasks.run_xstartup()
     # run_tasks.run_check_xorg()
     # run_tasks.run_test()   # Note, this is not paralleized
-    #
-    # # Command to actually run the tasks.
+
+    # Command to actually run the tasks.
     run_tasks.run_tasks()
