+++ pipeline/opics_task_runner.py
@@ -6,18 +6,19 @@ import threading
 from pipeline import logger
 from pipeline import util
 from pipeline.mcs_test_runner import McsTestRunner
-from pipeline.mess_config_change import MessConfigChange
-from pipeline.mess_singletask import MessSingleTask
+from pipeline.opics_config_change import OpicsConfigChange
+from pipeline.opics_singletask import OpicsSingleTask
 from pipeline.xserver_check import XServerCheck
 from pipeline.xserver_startup import XServerStartup
 
 # Uncomment one of the following.  single is for testing;  the other
 # is for intphys
 # TASK_FILE_PATH = "tasks_delta_echo_foxtrot.txt"
-TASK_FILE_PATH = "scenes_single_scene.txt"
+# TASK_FILE_PATH = "scenes_single_scene.txt"
+TASK_FILE_PATH = "../scenes_juliett.txt"
 
 
-class MessRunTasks:
+class OpicsRunTasks:
 
     def __init__(self):
         self.available_machines = []
@@ -48,7 +49,7 @@ class MessRunTasks:
             task_file = task_files_list.pop(0)
             lock.release()
 
-            singleTask = MessSingleTask(machine_dns, task_file, threadlog)
+            singleTask = OpicsSingleTask(machine_dns, task_file, threadlog)
             return_code = singleTask.process()
 
             if return_code > 0:
@@ -73,13 +74,6 @@ class MessRunTasks:
         self.log.info(f"Tasks {task_files_list}")
 
     def run_tasks(self):
-        self.get_tasks()
-
-        # Determine the DNS for all the machine that we have, default to
-        # us-east-1 and p2.xlarge
-        self.available_machines = util.get_aws_machines()
-        self.log.info(f"Machines available {self.available_machines}")
-
         # Create a thread for each machine
         threads = []
         for machine in self.available_machines:
@@ -95,15 +89,17 @@ class MessRunTasks:
         self.log.info("Ending runtasks")
 
     def get_machines(self):
-        self.available_machines = util.get_aws_machines()
+        self.available_machines = util.get_aws_machines(tag_name='ta1', tag_value='opics')
         self.log.info(f"Number of machines {len(self.available_machines)}")
         self.log.info(f"Machines available:  {self.available_machines}")
 
     def run_xstartup(self):
         ''' Start X Server on all the machines.  Note:  Not parallelized'''
         for machine in self.available_machines:
-            xserver = XServerStartup(machine, self.log)
-            xserver.process()
+            cmd = "sudo /usr/bin/Xorg :0"
+            return_code = util.shell_run_background_remote(machine, cmd, self.log)
+            if not return_code == 0:
+                self.log.warn(f"Error starting x on {machine}")
 
     def kill_and_restartX(self):
         for machine in self.available_machines:
@@ -112,35 +108,31 @@ class MessRunTasks:
 
     def change_mcs_config(self):
         for machine in self.available_machines:
-            config_change = MessConfigChange(machine, self.log)
+            config_change = OpicsConfigChange(machine, self.log)
             config_change.process()
 
     def run_check_xorg(self):
         ''' Check X Server on all the machines.  Note:  Not parallelized'''
-        self.available_machines = util.get_aws_machines()
-        self.log.info(f"Machines available {self.available_machines}")
-
         for machine in self.available_machines:
             xserver_check = XServerCheck(machine, self.log)
             xserver_check.process()
 
     def run_test(self):
-        self.available_machines = util.get_aws_machines()
         for machine in self.available_machines:
             test_runner = McsTestRunner(machine, self.log)
             test_runner.process()
 
 
 if __name__ == '__main__':
-    run_tasks = MessRunTasks()
+    run_tasks = OpicsRunTasks()
     run_tasks.get_machines()
     run_tasks.get_tasks()
 
     # Commands to change the Remote machines.  Uncomment them to run them.
     # run_tasks.change_mcs_config()
-    # run_tasks.runXStartup()
-    # run_tasks.runCheckXorg()
+    # run_tasks.run_xstartup()
+    # run_tasks.run_check_xorg()
     # run_tasks.run_test()   # Note, this is not paralleized
-
-    # Command to actually run the tasks.
+    #
+    # # Command to actually run the tasks.
     run_tasks.run_tasks()
